retrofit calls for 4 categories
