+++ app/src/main/java/com/android/tipsytrivia/MainActivity.java
@@ -1,5 +1,6 @@
 package com.android.tipsytrivia;
 
+import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
@@ -21,22 +22,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        TriviaRetrofit.getInstance()
-                .create(TipsyTriviaApi.class)
-                .getTrivia("14", "easy")
-                .enqueue(new Callback<TipsyResponse>() {
-                    @Override
-                    public void onResponse(Call<TipsyResponse> call, Response<TipsyResponse> response) {
-                        Log.d("joestag", "onResponse: " + response.body().getResults().get(0).getCategory());
 
-                    }
-
-                    @Override
-                    public void onFailure(Call<TipsyResponse> call, Throwable t) {
-
-                        Log.d("joestag", "onFailure: " + t.getMessage());
-                    }
-                });
