+++ app/src/main/java/com/android/tipsytrivia/SplashActivity.java
@@ -25,7 +25,7 @@ public class SplashActivity extends AppCompatActivity {
         runnable = new Runnable() {
             @Override
             public void run() {
-                startActivity(new Intent(getApplicationContext(), MainActivity.class));
+                startActivity(new Intent(getApplicationContext(), CategoryPageActivity.class));
                 finish();
             }
